refactor: move face tracker to controllers
