+++ Server/app/controllers/__init__.py
@@ -1,0 +1,6 @@
+"""Controller package for application components."""
+
+from .face_tracker import FaceTracker
+
+__all__ = ["FaceTracker"]
+
